+++ Reinvest/Trading/migrations/index.ts
@@ -1,0 +1,1 @@
+export * as migration_20230523143501_create_trades from './trading_1_create_trades';