+++ CHANGELOG.md
@@ -1,5 +1,29 @@
 # REINVEST API CHANGELOG
 
+## 1.14.0 - 05/23/2023
+
+* Implementing investment process manager
+* Investment process manager includes steps:
+    * VerifyAccountForInvestment: Verify account for investment (KYC, AML). Confirm if all were verified
+    * FinalizeInvestment: [MOCK] Generate subscription agreement, confirm fees
+    * CreateTrade:
+        * Create trade in North Capital
+        * Create drafted distribution in Vertalo
+        * Upload subscription agreement to North Capital trade
+        * Initiate funds transfer from external bank account to REINVEST escrow account
+        * Open Vertalo distribution (waiting for funds transfer)
+    * CheckIsInvestmentFunded:
+        * Check if funds transfer is completed in North Capital
+        * Close Vertalo distribution (funds transfer completed)
+    * CheckIsInvestmentApproved: [MOCK] Check trade RR Approval status by Dalmore
+    * CheckIsGracePeriodEnded: [MOCK] Check if grace period is ended
+    * MarkFundsAsReadyToDisburse: [MOCK] Mark funds as ready to disburse
+    * TransferSharesWhenTradeSettled: [MOCK]
+        * Verify if trade is settled
+        * Mark payment as completed in Vertalo
+        * Transfer shares from REINVEST to investor in Vertalo
+    * FinishInvestment: [MOCK] Finish investment
+
 ## 1.13 - 05/17/2023
 
 * Type `Money` from now on accepts only integer values instead of strings.